JWT and Authorization added

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalsModel");
+const User = require("../models/userModel");
 
 //@decription Get Goals
 //@route      GET /api/goals
 //@access     private
 const getGoals = asyncHandler(async(req,res) =>{
 
-    const allGoals = await Goal.find();
+    const allGoals = await Goal.find({ user : req.user.id });
 
     res.status(200).json(allGoals);
 })
@@ -23,6 +24,7 @@ const setGoal = asyncHandler(async(req,res) => {
 
     const goal = await Goal.create({
         title : req.body.title,
+        user : req.user.id,
     })
 
     res.status(201).json(goal);
@@ -39,6 +41,18 @@ const UpdateGoal = asyncHandler(async(req,res) => {
         throw new Error("Goal Not found")
     }
 
+    const user = await User.findById(req.user.id);
+    //Check the user
+    if(!user){
+        res.status(401)
+        throw new Error(" User not found");
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error(" User not Authorized");
+    }
+    
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body, { new : true});
     res.status(200).json(updatedGoal);
 })
@@ -52,6 +66,18 @@ const deleteGoal = asyncHandler(async(req,res) => {
     if(!deletedGoal){
         res.status(400)
         throw new Error("Goal Not found")
+    }
+
+    const user = await User.findById(req.user.id);
+    //Check the user
+    if(!user){
+        res.status(401)
+        throw new Error(" User not found");
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error(" User not Authorized");
     }
     
     await Goal.findByIdAndDelete(req.params.id);
